requirements.txt
# 아파트 매물관리 자동화 시스템 의존성 (Python 3.9 호환)
# Core dependencies
gspread>=5.0.0
pandas>=1.5.0,<2.1.0
numpy>=1.21.0,<1.25.0
requests>=2.28.0

# Google Services
google-auth>=2.15.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.80.0
google-generativeai>=0.3.0

# Data processing
pdfplumber>=0.7.0
openpyxl>=3.0.0
xlrd>=2.0.0
xlsxwriter>=3.0.0

# Web scraping & automation
playwright>=1.30.0,<1.41.0
selenium>=4.10.0

# Image processing
pillow>=9.0.0

# Configuration & utilities
python-dotenv>=0.19.0
pyyaml>=6.0.0
schedule>=1.1.0
loguru>=0.6.0
pydantic>=1.10.0,<2.0.0

# MCP Servers
mcp-google-sheets

# Development tools
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0
pre-commit>=3.0.0

# Optional: ML/AI
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
